+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+BUF FIXES
+* Fix issue where the `acl-controller` did not update the default namespace with the cross-namespace policy.
+  [[GH-104](https://github.com/hashicorp/consul-ecs/pull/104)]
+
 ## 0.5.0-beta1 (Jun 06, 2022)
 
 BREAKING CHANGES